+++ constraint_counter/src/noir.rs
@@ -1,123 +1,153 @@
-// use acvm::{
-//     acir::{
-//         acir_field::GenericFieldElement,
-//         circuit::{brillig::BrilligBytecode, Circuit, Program},
-//         native_types::{Witness as AcvmWitness, WitnessMap},
-//     },
-//     blackbox_solver::StubbedBlackBoxSolver,
-//     pwg::ACVM,
-// };
-// use ark_ff::PrimeField;
-// use ark_r1cs_std::{alloc::AllocVar, fields::fp::FpVar, R1CSVar};
-// use ark_relations::r1cs::{ConstraintSynthesizer, ConstraintSystemRef, SynthesisError};
+use std::collections::HashMap;
 
-// use serde::{self, Deserialize, Serialize};
-// use std::collections::HashMap;
+use acvm::{
+    acir::{
+        acir_field::GenericFieldElement,
+        circuit::{brillig::BrilligBytecode, Circuit, Opcode, Program},
+        native_types::{Witness, WitnessMap},
+    },
+    blackbox_solver::StubbedBlackBoxSolver,
+    pwg::ACVM,
+    AcirField,
+};
+use ark_ff::Field;
+use ark_relations::r1cs::{ConstraintSynthesizer, LinearCombination, Variable};
+use serde::{Deserialize, Serialize};
 
-// use super::*;
-// use crate::bridge::AcirCircuitSonobe;
+use super::*;
 
-// #[derive(Clone, Debug)]
-// pub struct NoirCircuit {
-//     pub circuit: Circuit<GenericFieldElement<Fr>>,
-//     pub unconstrained_functions: Vec<BrilligBytecode<GenericFieldElement<Fr>>>,
-// }
+#[derive(Clone, Serialize, Deserialize, Debug)]
+pub struct NoirProgram {
+    #[serde(
+        serialize_with = "Program::serialize_program_base64",
+        deserialize_with = "Program::deserialize_program_base64"
+    )]
+    pub bytecode: Program<GenericFieldElement<Fr>>,
+}
 
-// #[derive(Clone, Serialize, Deserialize, Debug)]
-// pub struct ProgramArtifactGeneric<F: PrimeField> {
-//     #[serde(
-//         serialize_with = "Program::serialize_program_base64",
-//         deserialize_with = "Program::deserialize_program_base64"
-//     )]
-//     pub bytecode: Program<GenericFieldElement<F>>,
-// }
+impl NoirProgram {
+    pub fn new(bin: &[u8]) -> Self {
+        serde_json::from_slice(bin).unwrap()
+    }
 
-// impl NoirCircuit {
-//     pub fn new(bin: &[u8]) -> Self {
-//         let program: ProgramArtifactGeneric<Fr> = serde_json::from_slice(bin).unwrap();
-//         let circuit: Circuit<GenericFieldElement<Fr>> = program.bytecode.functions[0].clone();
-//         // let public_io_length = circuit.public_parameters.0.len();
-//         // let ivc_return_length = circuit.return_values.0.len();
+    pub fn circuit(&self) -> &Circuit<GenericFieldElement<Fr>> {
+        &self.bytecode.functions[0]
+    }
 
-//         if public_io_length != ivc_return_length {
-//             panic!("Public input and output are not the same length!'\n    IVC input: {public_io_length}\n    IVC output: {ivc_return_length}");
-//         }
+    pub fn unconstrained_functions(&self) -> &Vec<BrilligBytecode<GenericFieldElement<Fr>>> {
+        &self.bytecode.unconstrained_functions
+    }
 
-//         Self {
-//             circuit,
-//             unconstrained_functions: program.bytecode.unconstrained_functions,
-//         }
-//     }
+    pub fn solve(
+        &self,
+        instance_variables: &[Fr],
+        witness_variables: &[Fr],
+    ) -> WitnessMap<GenericFieldElement<Fr>> {
+        let mut acvm = ACVM::new(
+            &StubbedBlackBoxSolver(false),
+            &self.circuit().opcodes,
+            WitnessMap::new(),
+            self.unconstrained_functions(),
+            &[],
+        );
 
-//     pub fn generate_step_constraints(
-//         &self,
-//         cs: ConstraintSystemRef<Fr>,
-//         z_i: &[FpVar<Fr>],
-//         external_inputs: &[FpVar<Fr>],
-//     ) -> Result<Vec<FpVar<Fr>>, SynthesisError> {
-//         let mut acvm = ACVM::new(
-//             &StubbedBlackBoxSolver(false),
-//             &self.circuit.opcodes,
-//             WitnessMap::new(),
-//             &self.unconstrained_functions,
-//             &[],
-//         );
+        self.circuit()
+            .public_parameters
+            .0
+            .iter()
+            .for_each(|witness| {
+                let f =
+                    GenericFieldElement::<Fr>::from_repr(instance_variables[witness.as_usize()]);
+                acvm.overwrite_witness(*witness, f);
+            });
 
-//         let mut already_assigned_witness_values = HashMap::new();
+        // write witness values for external_inputs
+        self.circuit()
+            .private_parameters
+            .iter()
+            .for_each(|witness| {
+                let idx = dbg!(witness.as_usize()) - dbg!(instance_variables.len());
 
-//         self.circuit.public_parameters.0.iter().for_each(|witness| {
-//             let idx: usize = witness.as_usize();
-//             let witness = AcvmWitness(witness.witness_index());
-//             already_assigned_witness_values.insert(witness, &z_i[idx]);
-//             let val = z_i[idx].value().unwrap();
+                let f = GenericFieldElement::<Fr>::from_repr(witness_variables[idx]);
+                acvm.overwrite_witness(*witness, f);
+            });
+        let _status = acvm.solve();
+        acvm.finalize()
+    }
+}
 
-//             let f = GenericFieldElement::<Fr>::from_repr(val);
-//             acvm.overwrite_witness(witness, f);
-//         });
+impl ConstraintSynthesizer<Fr> for NoirProgram {
+    fn generate_constraints(self, cs: ConstraintSystemRef<Fr>) -> ark_relations::r1cs::Result<()> {
+        let mut witness_map: HashMap<Witness, Variable> = HashMap::new();
 
-//         // write witness values for external_inputs
-//         self.circuit.private_parameters.iter().for_each(|witness| {
-//             let idx = witness.as_usize() - z_i.len();
-//             let witness = AcvmWitness(witness.witness_index());
-//             already_assigned_witness_values.insert(witness, &external_inputs[idx]);
+        // First, allocate all public inputs
+        let public_inputs = &self.circuit().public_inputs().0;
+        for &witness in public_inputs {
+            let var = cs.new_input_variable(|| Ok(Fr::ZERO))?;
+            witness_map.insert(witness, var);
+        }
 
-//             let val = external_inputs[idx].value().unwrap();
+        // Then, allocate known private witnesses
+        let private_params = &self.circuit().private_parameters;
+        for &witness in private_params {
+            let var = cs.new_witness_variable(|| Ok(Fr::ZERO))?;
+            witness_map.insert(witness, var);
+        }
 
-//             let f = GenericFieldElement::<Fr>::from_repr(val);
-//             acvm.overwrite_witness(witness, f);
-//         });
+        for opcode in &self.circuit().opcodes {
+            if let Opcode::AssertZero(gate) = opcode {
+                let mut left_terms = LinearCombination::<Fr>::new();
+                let mut right_terms = LinearCombination::<Fr>::new();
+                let mut output_terms = LinearCombination::<Fr>::new();
 
-//         // computes the witness
-//         let start = std::time::Instant::now();
-//         let _status = acvm.solve();
-//         let witness_map = acvm.finalize();
-//         println!("ACVM solve and finalize: {:?}", start.elapsed());
+                // Handle multiplication terms
+                for mul_term in &gate.mul_terms {
+                    let coeff = Fr::from(mul_term.0.into_repr());
+                    let left_var = allocate_variable(&mut witness_map, mul_term.1, &cs)?;
+                    let right_var = allocate_variable(&mut witness_map, mul_term.2, &cs)?;
 
-//         // get the z_{i+1} output state
-//         let assigned_z_i1 = self
-//             .circuit
-//             .return_values
-//             .0
-//             .iter()
-//             .map(|witness| {
-//                 let noir_field_element = witness_map
-//                     .get(witness)
-//                     .ok_or(SynthesisError::AssignmentMissing)?;
-//                 FpVar::<Fr>::new_witness(cs.clone(), || Ok(noir_field_element.into_repr()))
-//             })
-//             .collect::<Result<Vec<FpVar<Fr>>, SynthesisError>>()?;
+                    left_terms += (coeff, left_var);
+                    right_terms += (Fr::ONE, right_var);
+                }
 
-//         // initialize circuit and set already assigned values
-//         let mut acir_circuit = AcirCircuitSonobe::from((&self.circuit, witness_map));
-//         acir_circuit.already_assigned_witnesses = already_assigned_witness_values;
+                // Handle linear combinations (add terms)
+                for add_term in &gate.linear_combinations {
+                    let coeff = Fr::from(add_term.0.into_repr());
+                    let var = allocate_variable(&mut witness_map, add_term.1, &cs)?;
 
-//         let start = std::time::Instant::now();
-//         acir_circuit.generate_constraints(cs)?;
-//         println!(
-//             "Duration for `acir_circuit.generate_constraints(cs)`: {:?}",
-//             start.elapsed()
-//         );
+                    // Add to the output terms
+                    output_terms += (coeff, var);
+                }
 
-//         Ok(assigned_z_i1)
-//     }
-// }
+                // Add constant term if present
+                if !gate.q_c.is_zero() {
+                    output_terms += (Fr::from(gate.q_c.into_repr()), Variable::One);
+                }
+
+                // The constraint becomes: left_terms * right_terms + output_terms = 0
+                cs.enforce_constraint(left_terms, right_terms, -output_terms)?;
+            }
+            if let Opcode::MemoryInit { .. } | Opcode::MemoryOp { .. } = opcode {
+                panic!("Memory Opcode was used! This is not currently supported.");
+            }
+        }
+
+        Ok(())
+    }
+}
+
+// Helper function to allocate variables in the constraint system
+fn allocate_variable(
+    witness_map: &mut HashMap<Witness, Variable>,
+    witness: Witness,
+    cs: &ConstraintSystemRef<Fr>,
+) -> ark_relations::r1cs::Result<Variable> {
+    if let Some(&var) = witness_map.get(&witness) {
+        Ok(var)
+    } else {
+        // Create a new variable without needing concrete witness values
+        let var = cs.new_witness_variable(|| Ok(Fr::ZERO))?;
+        witness_map.insert(witness, var);
+        Ok(var)
+    }
+}
